Move helper fun to t4_lambda_shared
Makes testing easier.
